core + crypto + io: switch length params to int64_t
Consistency pass over emp-tool's public API. Length-typed parameters
(buffer counts, byte counts, IOChannel send/recv sizes) used a mix of
`int` and `size_t`; both have footguns:

- `int` silently overflows at 2^31 elements. Today's chunks stay well
  below that, but a callsite that multiplies (e.g. `nblocks * 128` for
  a bit count) can overflow at much smaller `nblocks`.
- `size_t` underflows silently in `while (len > 0) { ...; len -= batch; }`
  patterns — a pre-condition mistake on `batch` turns into a 2^64-block
  loop instead of a bounds error.

`int64_t` avoids both: signed 64-bit, no underflow, range that vastly
exceeds anything plausibly in scope. AGENTS.md gains a top-level rule
documenting the convention for future contributors.

Widened (`int` → `int64_t`):
- core/block: xorBlocks_arr (3 overloads), xorBlocksTo_arr, cmpBlock.
- core/utils: bools_to_bits, bits_to_bools.
- crypto/prg: random_data, random_data_unaligned, random_bool, random_block.
- crypto/prp: permute_block.
- crypto/aes: ParaEnc(K, N), ParaCtrEnc(n), AES_ecb_encrypt_blks(nblks).
- crypto/ccrh: Hn.
- crypto/hash: put, put_block, hash_once, hash_for_block.
- crypto/f2k: vector_inn_prdt_sum_no_red, vector_inn_prdt_sum_red (2
  overloads), uni_hash_coeff_gen, vector_self_xor, inn_prdt_drain_lanes.

Widened (`size_t` → `int64_t`):
- io/io_channel: virtual send_data_internal / recv_data_internal, plus
  every helper (send_data, recv_data, send_block, recv_block, send_pt,
  recv_pt, send_bool, recv_bool).
- io/net_io_channel + io/tls_io_channel + io/trace_io: every override
  of the two virtuals.
- test/test_garble: AbandonIO's override pair.

Function bodies' local counters / loop indices follow suit where they
participated in the widened-param arithmetic.

Template non-type parameters (`int N`, `int K` in ParaEnc<>) stay as
`int` — compile-time bounded small constants, no overflow concern.

All 25 emp-tool tests pass.

+++ emp-tool/core/block.h
@@ -30,20 +30,20 @@ std::ostream& operator<<(std::ostream& out, const block& blk);
 // res[i] = x[i] ^ y[i]. Pointers must be pairwise non-overlapping.
 inline void xorBlocks_arr(block* __restrict__ res,
                           const block* __restrict__ x,
-                          const block* __restrict__ y, int nblocks);
+                          const block* __restrict__ y, int64_t nblocks);
 
 // In-place: dst[i] ^= src[i]. dst and src must not overlap.
 inline void xorBlocksTo_arr(block* __restrict__ dst,
-                            const block* __restrict__ src, int nblocks);
+                            const block* __restrict__ src, int64_t nblocks);
 
 // res[i] = x[i] ^ y (broadcast). res and x must not overlap.
 inline void xorBlocks_arr(block* __restrict__ res,
                           const block* __restrict__ x,
-                          block y, int nblocks);
+                          block y, int64_t nblocks);
 
 // True iff x[0..n) == y[0..n) element-wise. No early exit:
 // accumulator-OR exposes ILP on the equal-case path.
-inline bool cmpBlock(const block* x, const block* y, int nblocks);
+inline bool cmpBlock(const block* x, const block* y, int64_t nblocks);
 
 // --- SSE bit-matrix transpose --------------------------------------------
